feat: user login

+++ src/services/article/delete_article_service.js
@@ -1,6 +1,9 @@
 import prisma from "../prisma.js";
 
-export async function deleteArticleService(id) {
+export async function deleteArticleService({ id, user }) {
+  const article = await prisma.article.findUnique({ where: { id } });
+  if (!article) throw new Error("NOT_FOUND");
+  if (article.userId !== user.id) throw new Error("FORBIDDEN");
   await prisma.article.delete({
     where: { id },
   });
