+++ src/services/product/delete_product_service.js
@@ -1,6 +1,9 @@
 import prisma from "../prisma.js";
 
-export async function deleteProductService(id) {
+export async function deleteProductService({ id, user }) {
+  const product = await prisma.product.findUnique({ where: { id } });
+  if (!product) throw new Error("NOT_FOUND");
+  if (product.userId !== user.id) throw new Error("FORBIDDEN");
   await prisma.product.delete({
     where: { id },
   });
